[diagnostic_aggregator] fix matches not being found when number of analyzers changes
Diagnostics with names which should be matched would be matched by the Other
group due to the setup of the match function. Resetting the map means that all
match arrays are reinitialised whenever an analyzer is added or removed, which
ensures that the arrays are consistent with the current analyzers.