Give the bonds a bit longer to form in tests
This was failing on our jenkins as it was too short.

As it is, it's a bit unreliable for a test as the bonds may not have yet been formed but a new message with the diagnostics in the 'Other' can arrive and this will fail the test. Might not be worth doing anything complicated though if a reasonable timeout does the job.